Some methods moved in CEUtils
