+++ src/main/java/codeemoji/inlay/external/BaseExternalServiceForVulnerability.java
@@ -1,6 +1,7 @@
 package codeemoji.inlay.external;
 
 import codeemoji.core.external.CEExternalService;
+import codeemoji.core.util.CEUtils;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.roots.libraries.Library;
 import com.intellij.openapi.roots.libraries.LibraryTable;
@@ -15,7 +16,6 @@ import java.util.*;
 public abstract class BaseExternalServiceForVulnerability implements CEExternalService<VirtualFile, Object> {
     protected final Map<VirtualFile, Object> persistedData = new HashMap<>();
     protected Map<DependencyInfo, List<VulnerabilityInfo>> vulnerabilityMap = new HashMap<>();
-    protected final DependencyExtractor dependencyExtractor = new DependencyExtractor();
     protected DependencyInfo[] dependencyInfos;
 
     protected abstract VulnerabilityScanner getVulnerabilityScanner();
@@ -33,7 +33,7 @@ public abstract class BaseExternalServiceForVulnerability implements CEExternalS
         return Arrays.stream(librariesList)
                 .map(dep -> {
                     try {
-                        return dependencyExtractor.getDependecyInfo(dep);
+                        return CEUtils.getDependecyInfo(dep);
                     } catch (IllegalArgumentException e) {
                         System.err.println("Error processing dependency: " + dep.getName() + ". Error: " + e.getMessage());
                         return null;
